Check there are three instances on instance group
